Add style for sign-up button
Moves common text color to its own class. The header logo, hero main
text, and sign-up button all have the same color. I think that if we
were to change this color, we would want to change it on all of them.
Moves color to a common class so we would only have to change it in one
place.