Adjust call to action layout
Should have waited until I had formatted the text to do the layout.
Updates layout so that the call to action bubble will be the same width
as the rest of the page content by adjusting the call-to-action
container's padding. Adjust relative position of sign up button and text
to use flex box algorithm instead of hardcoding a gap.